Update morseTeensyLC_Keyboard.ino
Connected the LEDs and got them mostly where I want them.  Added, tested, and debugged <CR> and backspace support.
Added a partial powerup LED test. 
Need:
Clean up the LED reads and writes.  The phys IO should be done in ONE place, not all over the place.  A bitmap with the bits being an array index and the array elements being the pin to turn on/off is probably the cleanest answer.  Or-equal the turnON and turnOFF commands and arbitrate them once seems like the right way.